Add files via upload
Laptop Repair Services Management System for a shop in Kuala Lumpur.  This  system  is  divided  into  three  components:  one  for  administration,  receptionist, technician,  and  customer.  Every  subsystem  needs  a  log  in  function,  username  and  password  is there to prevent outsiders from accessing information, which prohibits unauthorized access to the system.